pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Maicol Rodrigues"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.13.3"
fastapi = "^0.110.0"
chromadb = "^0.4.24"
pandas = "^2.2.1"
sentence-transformers = "^2.5.1"
pydantic = "^2.6.3"
openpyxl = "^3.1.2"
pandera = "^0.18.0"
tiktoken = "^0.6.0"
psycopg = "^3.1.18"
langchain-community = "^0.0.28"
langchain = "^0.1.12"
langchain-openai = "^0.0.8"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
xlrd = "^2.0.1"
pyjwt = "^2.8.0"
cryptography = "^42.0.5"
boto3 = "^1.34.140"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
